change the input/output dialogs to allow direct editing
Forcing the path to change only through the folder dialog is very annoying and the dialog itself is clunky with no easy replacements. Simply changing the ReadOnly bool to false allows the user to edit this at their whim and has no issues, as long as the text results in a valid path.