updating the shader and surfaceproperties dropdown to include everything listed in the valve developer wiki
Currently it only shows as what I assume is the most commonly used Shaders and Surfaceproperties, or what was listed at the time of the creation of this list. Either way I'm updating it to include everything (except some surfaceproperties that have a note that says it's possibly missing), because why not and I don't need to look at the wiki to see all the accepted values.