experimental UI changes
Removed the Unused and Alpha flags from the checkbox list. Adjusted the labels to not get text wrapped, and tidied some other elements. Made the Frame, Face, Slice, and Mipmap number boxes in the Image tab set to ReadOnly if the value does not go over 1 (in this case, 0 = 1 as the starting point), this is to make it easier to see if the VTF has any more than one of any of these values at a glance instead of needing to attempt to change the values (they cannot be changed if it does not go over one as well, so this does not remove functionality)